+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasetitemop.py
@@ -1100,5 +1100,5 @@ class CreateDatasetItemResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-08-07T14:39:06.330Z")
+    updated: Optional[datetime] = parse_datetime("2025-08-08T08:03:31.826Z")
     r"""The date and time the resource was last updated"""