+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -100,7 +100,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K22FHKDDFKF1FFMSV4XHXD5G"
+        "01K24B9ZS35EH8WDN5CGS5TDBR"
     )
     r"""The id of the resource"""