+++ packages/orq-rc/RELEASES.md
@@ -2199,3 +2199,13 @@ Based on:
 - [python v3.11.0-rc.2] packages/orq-rc
 ### Releases
 - [PyPI v3.11.0-rc.2] https://pypi.org/project/orq-ai-sdk/3.11.0-rc.2 - packages/orq-rc
+
+## 2025-08-08 08:04:58
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.599.0 (2.675.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.11.0-rc.3] packages/orq-rc
+### Releases
+- [PyPI v3.11.0-rc.3] https://pypi.org/project/orq-ai-sdk/3.11.0-rc.3 - packages/orq-rc